Cleanup of legacy code, Removed old CurrentAccountMenu
